[core][Android] Rename `ExperimentalJSTypeConverter` -> `JSTypeConverter`
