+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/jni/JavaScriptFunction.kt
@@ -3,7 +3,7 @@ package expo.modules.kotlin.jni
 import com.facebook.jni.HybridData
 import expo.modules.core.interfaces.DoNotStrip
 import expo.modules.kotlin.AppContext
-import expo.modules.kotlin.types.JSTypeConverter
+import expo.modules.kotlin.types.JSTypeConverterProvider
 import expo.modules.kotlin.types.LazyKType
 import expo.modules.kotlin.types.TypeConverterProviderImpl
 import kotlin.reflect.KType
@@ -22,7 +22,7 @@ class JavaScriptFunction<ReturnType : Any?> @DoNotStrip private constructor(@DoN
   operator fun invoke(vararg args: Any?, thisValue: JavaScriptObject? = null, appContext: AppContext? = null): ReturnType {
     // TODO(@lukmccall): check current thread
     val convertedArgs = args
-      .map { JSTypeConverter.convertToJSValue(it) }
+      .map { JSTypeConverterProvider.convertToJSValue(it) }
       .toTypedArray()
 
     val converter = TypeConverterProviderImpl
