+++ packages/expo-modules-test-core/android/src/main/java/expo/modules/test/core/legacy/TestJSContainerProvider.kt
@@ -4,9 +4,9 @@ import com.facebook.react.bridge.JavaOnlyArray
 import com.facebook.react.bridge.JavaOnlyMap
 import com.facebook.react.bridge.WritableArray
 import com.facebook.react.bridge.WritableMap
-import expo.modules.kotlin.types.JSTypeConverter
+import expo.modules.kotlin.types.JSTypeConverterProvider
 
-internal object TestJSContainerProvider : JSTypeConverter.ContainerProvider {
+internal object TestJSContainerProvider : JSTypeConverterProvider.ContainerProvider {
   override fun createMap(): WritableMap = JavaOnlyMap()
   override fun createArray(): WritableArray = JavaOnlyArray()
 }
